move RepositoryViewer components into own folder
